Refactoring of the tests folder
