+++ cli/tox.ini
@@ -17,7 +17,7 @@ commands =
     # Running with discover and not unittest discover for Python 2.6 compatibility
     python -m discover -s tests/pcluster -p "*_test.py"
     # awsbatch-cli is not currently compatible with Python2.6
-    py{27,34,35,36,37}: py.test -l -v --basetemp={envtmpdir} --html=report.html --cov={envsitepackagesdir}/awsbatch tests/awsbatch/
+    py{27,34,35,36,37}: py.test -l -v --basetemp={envtmpdir} --html=report.html --cov={envsitepackagesdir}/awsbatch --cov={envsitepackagesdir}/pcluster tests
 
 # Section used to define common variables used by multiple testenvs.
 [vars]
